requirements.txt
Flask==3.0.0
opencv-python==4.10.0.84
numpy==1.26.4
Pillow==10.3.0
requests==2.32.3
